+++ src/main/java/com/benahmed/gestiondestock/service/impl/ClientServiceImpl.java
@@ -4,8 +4,10 @@ import com.benahmed.gestiondestock.DTO.ClientDto;
 import com.benahmed.gestiondestock.exception.EntityNotFoundException;
 import com.benahmed.gestiondestock.exception.ErrorCodes;
 import com.benahmed.gestiondestock.exception.InvalidEntityException;
-import com.benahmed.gestiondestock.model.Client;
+import com.benahmed.gestiondestock.exception.InvalidOperationException;
+import com.benahmed.gestiondestock.model.CommandeClient;
 import com.benahmed.gestiondestock.repository.ClientRepository;
+import com.benahmed.gestiondestock.repository.CommandeClienttRepository;
 import com.benahmed.gestiondestock.service.ClientService;
 import com.benahmed.gestiondestock.validator.ClientValidator;
 import lombok.extern.slf4j.Slf4j;
@@ -13,17 +15,18 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
 @Slf4j
 public class ClientServiceImpl implements ClientService {
 
-    ClientRepository clientRepository;
+    private final ClientRepository clientRepository;
+    private final CommandeClienttRepository commandeClienttRepository;
     @Autowired
-    public ClientServiceImpl(ClientRepository clientRepository) {
+    public ClientServiceImpl(ClientRepository clientRepository, CommandeClienttRepository commandeClienttRepository) {
         this.clientRepository = clientRepository;
+        this.commandeClienttRepository = commandeClienttRepository;
     }
 
     @Override
@@ -65,6 +68,11 @@ public class ClientServiceImpl implements ClientService {
         if (id == null){
             log.error("l'id est null");
             return;
+        }
+        List<CommandeClient> commandeClients = commandeClienttRepository.findAllByClientId(id);
+        if(!commandeClients.isEmpty()){
+            throw new InvalidOperationException("impossible de supprimer un client qui a deja des commandes clients",
+                    ErrorCodes.CLIENT_ALREADY_IN_USE);
         }
         clientRepository.deleteById(id);
     }
